improvement(landing): extract shared LandingField component
